Made gear slots always visible
Now the gear slots move to the bottom left corner when you don't have
the inventory. Otherwise, they move back to the top of the inventory.
They are only able to be edited if you have the inventory opened though.
This also allows for the oxygen on scuba tanks to be monitored in real
time along with durability to be monitored as well.