Durabilities for all equipment
Now all gear have durability, and of course, when the durability runs
out, the gear disappears. Walking with shoes and flippers damage them.
Also the damage increases depending on the type of ground your walking
on and the shoes/flippers you're wearing. Scuba tanks lose durability
from refilling with bubbles.