Equipment Commitment
I have made adding items into the game so easy! In the last 30 minutes I
added in Snorkels, Flip Flops, and Water Shoes! A lot of which are
Normal loot category because the only loot pool that was getting any new
items was the Major pool. The rest just had 1. :P I think I'll load the
game up with potions now!